% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-summarize.R
\name{agg}
\alias{agg}
\title{Aggregation}
\usage{
agg(aggregation, ...)
}
\arguments{
\item{aggregation}{one of sum, mean, or similar function}
}
\description{
Aggregation functions for grouped data.frames
}
