% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{chart_viz_options}
\alias{chart_viz_options}
\alias{dsviz_default_opts}
\title{Chart Options}
\usage{
dsviz_default_opts(...)
}
\arguments{
\item{title}{String with the main title (on top). Default is NULL}

\item{subtitle}{String for subtitle (below the title). Default is NULL}

\item{caption}{String for caption (bottom). Default is NULL}

\item{hor_title}{String. The main title (on top). Default is NULL}

\item{ver_title}{String. The main title (on top). Default is NULL}

\item{legend_title}{String. The main title (on top). Default is NULL}

\item{palette_colors}{An optional character vector specifying the colors}

\item{palette_type}{String with type of color palette (optional). You can choose into
"categorical", "sequential" and "divergent" palettes, more info in  <\code{\link[=dsvizopts_palette_type]{palette-types}}>}

\item{color_by}{A character with the name of the variable by which you want to color the graph. Default is NULL}

\item{drop_na}{Boolean. If TRUE drop missing values}

\item{na_label}{String with label that should be used to recode missing values. Default is "(NA)"}

\item{drop_na_legend}{Boolean. If TRUE drop missing values in legend}

\item{agg}{Compute summary statistics of data, the default aggregation is "sum" but "mean" and "median" are available}

\item{agg_text}{When there is aggregation agg_text is prepended to grouped variable}

\item{sort}{String. Order the bars by the values of the graphed category. Options: "without", "desc" (sort a variable in descending order), "asc" (sort a variable in ascending order). Default is "without.}

\item{slice_n}{Numeric value, It allows you to select and remove unique rows}

\item{order}{A character vector of variables to sort categories on the axes in the respective order.}

\item{order_legend}{A character vector of variables to sort categories on the legend in the respective order.}

\item{percentage}{Boolean. If TRUE, calculates the percentage according to the variable chosen in "percetage_col", by default it chooses the first column of the entered dataframe.}

\item{percentage_col}{A character with the name of the variable for which you want the percentage to be calculated.}

\item{format_sample_cat}{A specifier to format categories, e.g. "TITTLE", "lowercase", "Title Case"}

\item{format_sample_num}{A specifier to format numerical value(s), e.g. "1,234.56" or "1'234.5678"}

\item{format_sample_dat}{A specifier to format date(s), e.g. "28/04/2020" or "February 28th 2020"}

\item{locale}{Locale to use, e.g. "fr-FR" for french.}

\item{label_wrap}{A numeric vector of length one to choose the number of characters of the label text (overflowing text will be placed in a new line below).}

\item{label_wrap_legend}{A numeric vector of length one to choose the number of characters in each line of text in the legend (overflowing text will be placed in a new line below.)}

\item{spline}{Boolean. If TRUE, the segments between the data points are smoothed.}

\item{suffix}{Character string to append before formatted value.}

\item{prefix}{Character string to append after formatted value.}

\item{orientation}{A character vector of length one, where the chart can be displayed vertically ("ver") or horizontally ("hor").}

\item{graph_type}{A character vector of length one with graph type. Options: "grouped" (show individual values) or "stacked" (stack the values of each series on top of each other). Default is "grouped".}

\item{highlight_value}{A character vector with the categories you want to highlight. Default is NULL}

\item{highlight_value_color}{A character vector of length one with the color to highlight categories select in highlight_value}

\item{tooltip}{HTML with information that appears when hovering your pointer over the graph. Variables are enclosed by curly brackets.}

\item{allow_point}{Boolean. If TRUE, allow this series' points to be selected by clicking on the graphic. Default is FALSE.}

\item{color_click}{String with color to change the color of the series that was clicked.}

\item{color_hover}{String with color to change the color of the series over which the mouse hovered.}

\item{cursor}{You can set the cursor to "pointer" if you have click events attached to the series, to signal to the user that the points and lines can be clicked.}

\item{clickFunction}{String with JS function to indicate the values to be captured when clicking on the chart.}

\item{dataLabels_show}{Boolean. If TRUE it shows the values of the chart labels}

\item{dataLabels_format_sample}{String. A specifier to format label(s)}

\item{dataLabels_size}{Number. Label font size}

\item{dataLabels_color}{String with hexadecimal color for the label}

\item{dataLabels_text_outline}{Boolean. If TRUE, data labels are show with outline}

\item{opts}{All of the previous options can be passed as list of options opts}
}
\description{
This page describes the chart options arguments
}
