% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-map-prep.R
\name{data_map_prep}
\alias{data_map_prep}
\title{General function for processing geographic data}
\usage{
data_map_prep(
  data,
  ftype,
  agg,
  ptage_col = NULL,
  more_levels = FALSE,
  group_extra_num = TRUE
)
}
\arguments{
\item{data}{A data frame with geografical info.}

\item{ftype}{A string value with type of data to be plotted}

\item{agg}{Statistics which can be applied to all data subsets (sum, mean, median)}

\item{ptage_col}{A string value with the name of the categorical variable against which the percentage is calculated.}

\item{more_levels}{A logical indicating if the map has more than one territorial level.}

\item{group_extra_num}{A logical indicating}
}
\description{
General function for processing geographic data
}
\examples{

df <- sample_data("Gnm-Cat-Num")
data_map_prep(data = df, ftype = "Gnm-Cat-Num". agg = "sum")

df <- data.frame(id = "COL", "URY", "ARG")
df$id <- as_Gnm(df$id)
data_map_prep(data = df, ftype = "Gcd". agg = "mean")

df <- sample_data("Gnm-Gcd-Num-Cat-Cat-Num")
data_map_prep(data = df, ftype = "Gnm-Num", agg = "sum", more_levels = TRUE)

}
