% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{default_theme_opts}
\alias{default_theme_opts}
\title{Theme Options}
\usage{
default_theme_opts(
  logo = NULL,
  logo_position = NULL,
  palette_colors = NULL,
  palette_colors_categorical = NULL,
  palette_colors_sequential = NULL,
  palette_colors_divergent = NULL,
  palette_type = NULL,
  branding_include = NULL,
  branding_text = NULL,
  background_color = NULL,
  branding_background_color = NULL,
  accent_color = NULL,
  text_show = NULL,
  text_size = NULL,
  text_color = NULL,
  text_family = NULL,
  line_color = NULL,
  line_size = NULL,
  title_color = NULL,
  title_align = NULL,
  title_family = NULL,
  title_size = NULL,
  title_weight = NULL,
  subtitle_color = NULL,
  subtitle_align = NULL,
  subtitle_size = NULL,
  subtitle_weight = NULL,
  caption_color = NULL,
  caption_align = NULL,
  caption_size = NULL,
  axis_title_color = NULL,
  axis_title_size = NULL,
  axis_line_color = NULL,
  axis_line_size = NULL,
  axis_ticks_color = NULL,
  axis_tick_length = NULL,
  grid_x_enabled = NULL,
  grid_x_color = NULL,
  grid_x_size = NULL,
  grid_y_enabled = NULL,
  grid_y_color = NULL,
  grid_y_size = NULL,
  grid_y_dash = NULL,
  plot_margin_bottom = NULL,
  plot_margin_top = NULL,
  plot_margin_left = NULL,
  plot_margin_right = NULL,
  plot_background_color = NULL,
  plot_border_color = NULL,
  plot_border_size = NULL,
  label_color = NULL,
  label_size = NULL,
  label_text_decoration = NULL,
  label_text_outline = NULL,
  marker_enabled = TRUE
)
}
\arguments{
\item{logo}{Logo path or url}

\item{logo_position}{Logo "left" or "right"}

\item{palette_colors}{Vector of colors}

\item{palette_colors_categorical}{Vector of colors suitable for categorical variable}

\item{palette_colors_sequential}{Vector of sequential colors suitable for numerical variable}

\item{palette_colors_divergent}{Vector of divergent colors suitable for numerical variable}

\item{palette_type}{Type of color palette "categorical", "sequential", or "divergent"}

\item{branding_include}{Include branding logical}

\item{branding_text}{Text for branding bar}

\item{background_color}{Plot background color}

\item{branding_background_color}{Branding bar background color}

\item{accent_color}{?}

\item{text_show}{?}

\item{text_size}{Size of text in pt}

\item{text_color}{Text color}

\item{text_family}{Font}

\item{line_color}{Axis and grid line colors}

\item{line_size}{Line size}

\item{title_color}{Color for title text}

\item{title_align}{Alignment of title "left", "center" or "right"}

\item{title_family}{Title font}

\item{title_size}{Title size, defaults to 1.1 text_size}

\item{title_weight}{Font weight for title}

\item{subtitle_color}{Color for subtitle text}

\item{subtitle_align}{Alignment of subtitle "left", "center" or "right"}

\item{subtitle_size}{Subtitle size, defaults to 1.1 text_size}

\item{subtitle_weight}{Font weight for subtitle}

\item{caption_color}{Color for caption text}

\item{caption_align}{Alignment of caption "left", "center" or "right"}

\item{caption_size}{caption size, defaults to 1.1 text_size}

\item{axis_title_color}{Color for axis text}

\item{axis_title_size}{Axis text size}

\item{axis_line_color}{Axis line color}

\item{axis_line_size}{Axis line width}

\item{axis_ticks_color}{Axis ticks color}

\item{axis_tick_length}{Axis tick length}

\item{grid_x_enabled}{Enable horizontal grid}

\item{grid_x_color}{Horizontal grid color}

\item{grid_x_size}{Horizontal grid line width}

\item{grid_y_enabled}{Enable vertical grid}

\item{grid_y_color}{Vertical grid color}

\item{grid_y_size}{Verical grid line width}

\item{grid_y_dash}{?}

\item{plot_margin_bottom}{Plot margin bottom}

\item{plot_margin_top}{Plot margin top}

\item{plot_margin_left}{Plot margin left}

\item{plot_margin_right}{Plot margin right}

\item{plot_background_color}{? Chart background color}

\item{plot_border_color}{? Chart border color}

\item{plot_border_size}{? Chart border width}

\item{label_color}{? Text color}

\item{label_size}{? Text size}

\item{label_text_decoration}{? Text decoration}

\item{label_text_outline}{? Text outline}
}
\description{
Theme Options
}
