% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{dsviz_default_opts}
\alias{dsviz_default_opts}
\title{Chart Options}
\usage{
dsviz_default_opts(
  drop_na = NULL,
  na_label = NULL,
  agg = NULL,
  agg_text = NULL,
  sort = NULL,
  slice_n = NULL,
  order = NULL,
  axis_text_angle = NULL,
  color_by = NULL,
  format_dat_sample = NULL,
  format_num_sample = NULL,
  format_cat_sample = NULL,
  locale = NULL,
  label_wrap = NULL,
  spline = NULL,
  prefix = NULL,
  suffix = NULL,
  label_wrap_legend = NULL,
  orientation = NULL,
  bubble_min = NULL,
  bubble_max = NULL,
  data_labels_show = NULL,
  data_labels_format_sample = NULL,
  title = NULL,
  ver_title = NULL,
  hor_title = NULL,
  subtitle = NULL,
  caption = NULL,
  opts = NULL
)
}
\arguments{
\item{drop_na}{drop NA values from input data}

\item{na_label}{replace NA values with string}

\item{agg}{Aggregation function for groups, defaults to sum}

\item{agg_text}{When there is aggegregaton agg_text is prepended to grouped variable}

\item{sort}{Whether to sort the data column: asc or desc}

\item{slice_n}{Slide the data from 1 to n}

\item{order}{Give the order of the categories}

\item{axis_text_angle}{Rotate axis ticks test}

\item{color_by}{Column to map colors}

\item{format_dat_sample}{Sample format for dates, e.g. "28/04/2020" or "February 28th 2020"}

\item{format_num_sample}{Sample format for numbers, e.g. "1,234.56" or "1'234.5678"}

\item{format_cat_sample}{Sample format for categories, e.g. "TITTLE", "lowercase", "Title Case"}

\item{locale}{Locale to use, e.g. "en-US"}

\item{label_wrap}{Number of characters to wrap long tick titles}

\item{spline}{Spline TRUE or FALSE for line charts (html only)}

\item{prefix}{Add preffix to values}

\item{suffix}{Add suffix to values}

\item{label_wrap_legend}{Wrap the legend string the given number of characters}

\item{orientation}{Chart orientation: horizonal ("hor") or verical ("ver")}

\item{bubble_min}{Relative min size of bubbles in chart}

\item{bubble_max}{Relative max size of bubbles in chart}

\item{data_labels_show}{Show data labels}

\item{data_labels_format_sample}{Format for data labels}

\item{title}{Chart title}

\item{ver_title}{Vertical axis title}

\item{hor_title}{Horizontal axis titble}

\item{subtitle}{Subtitle}

\item{caption}{Caption}

\item{opts}{All of the previous options can be passed as list of options opts}
}
\description{
Chart Options
}
