% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-summarize.R
\name{function_agg}
\alias{function_agg}
\title{Sum, mean. median o count of initial data}
\usage{
function_agg(df, agg, to_agg, ...)
}
\description{
\code{function_agg} returns the operation of all the values present in its arguments.
}
\details{
This is a generic function: methods can be defined for it directly or via the
\code{\link{across}} group generic. For this to work properly, the arguments
\code{...} should be unnamed, and dispatch is on the first argument.
}
