% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelsXY.R
\name{labelsXY}
\alias{labelsXY}
\title{Orientation}
\usage{
labelsXY(hor_title, ver_title, nms, orientation)
}
\arguments{
\item{hor_title}{Horizontal title}

\item{ver_title}{Vertical title}

\item{nms}{Column names}

\item{orientation}{Plot orientation}
}
\value{
vector for x,y axis titles
}
\description{
Returns horLabel and verLabel in a vector depending orientation
}
