% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-map-prep.R
\name{shape_info}
\alias{shape_info}
\title{shape info}
\usage{
shape_info(map_name, ftype, by_col = "name", addRds = FALSE)
}
\arguments{
\item{map_name}{Map name}

\item{ftype}{Data class by column}

\item{by_col}{Variables to join centroids with shape file}
}
\description{
shape info
}
\examples{

shape_info("col_departments", ftype = "Gnm-Num")
}
