% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-summarize.R
\name{summarizeData}
\alias{summarizeData}
\title{Summarize Data}
\usage{
summarizeData(df, agg, to_agg, ...)
}
\arguments{
\item{df}{A data.frame}

\item{agg}{one of sum, mean, or similar function}

\item{to_agg}{Columns to aggregate}

\item{...}{Variables to group}
}
\description{
Summarise a data.frame with the given aggregation function
}
