% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{chart_viz_options}
\alias{chart_viz_options}
\alias{dsviz_default_opts}
\title{Chart Options}
\usage{
dsviz_default_opts(...)
}
\arguments{
\item{title}{String with the main title (on top). Default is NULL}

\item{subtitle}{String for subtitle (below the title). Default is NULL}

\item{caption}{String for caption (bottom). Default is NULL}

\item{hor_title}{String with horizontal axis title. Default is NULL}

\item{ver_title}{String with vertical axis title . Default is NULLL}

\item{legend_title}{String with the main title to legend. Default is NULL}

\item{palette_colors}{An optional character vector specifying the colors}

\item{palette_type}{String with type of color palette (optional). Options:
"categorical", "sequential" and "divergent" palettes, more info in  <\code{\link[=dsvizopts_palette_type]{palette-types}}>}

\item{color_by}{A character with the name of the variable by which you want to color the graph. Default is NULL}

\item{drop_na}{Boolean. If TRUE drop missing values}

\item{na_label}{String with label that should be used to recode missing values. Default is "(NA)"}

\item{drop_na_legend}{Boolean. If TRUE drop missing values in legend}

\item{agg}{Aggregation type to compute summary statistics of data. Options: "sum", "mean", "median". Default is "sum"}

\item{agg_text}{When there is aggregation agg_text is prepended to grouped variable}

\item{sort}{String. Order the bars by the values of the graphed category. Options: "without", "desc" (sort a variable in descending order), "asc" (sort a variable in ascending order). Default is "without.}

\item{slice_n}{Numeric value, It allows you to select and remove unique rows}

\item{order}{A character vector of variables to sort categories on the axes in the respective order.}

\item{order_legend}{A character vector of variables to sort categories on the legend in the respective order.}

\item{percentage}{Boolean. If TRUE, calculates the percentage according to the variable chosen in "percetage_col", by default it chooses the first column of the entered dataframe.}

\item{percentage_col}{A character with the name of the variable for which you want the percentage to be calculated.}

\item{format_sample_cat}{A specifier to format categories, e.g. "TITTLE", "lowercase", "Title Case"}

\item{format_sample_num}{A specifier to format numerical value(s), e.g. "1,234.56" or "1'234.5678"}

\item{format_sample_dat}{A specifier to format date(s), e.g. "28/04/2020" or "February 28th 2020"}

\item{locale}{Locale to use, e.g. "fr-FR" for french.}

\item{label_wrap}{Is this a numeric vector of length one or just an integer to choose the number of characters of the label text (overflowing text will be placed in a new line below).}

\item{label_wrap_legendIs}{this a numeric vector of length one or just an integer to choose the number of characters in each line of text in the legend (overflowing text will be placed in a new line below.)}

\item{spline}{Boolean. If TRUE, the segments between the data points are smoothed.}

\item{suffix}{Character string to append before formatted value.}

\item{prefix}{Character string to append after formatted value.}

\item{orientation}{A character vector of length one, where the chart can be displayed vertically ("ver") or horizontally ("hor").}

\item{graph_type}{A character vector of length one with graph type. Options: "grouped" (show individual values) or "stacked" (stack the values of each series on top of each other). Default is "grouped".}

\item{highlight_value}{A character vector with the categories you want to highlight. Default is NULL}

\item{highlight_value_color}{A character vector of length one with the color to highlight categories select in highlight_value}

\item{tooltip}{HTML with information that appears when hovering your pointer over the graph. Variables are enclosed by curly brackets.}

\item{groupBy}{Character vector of column names to group by.}

\item{sortable}{Enable sorting? Default is TRUE.}

\item{resizable}{Enable column resizing? Default is FALSE.}

\item{filterable}{Enable column filtering? Default is FALSE.}

\item{searchable}{Enable global table searching? Default is FALSE.}

\item{selection}{Enable row selection? Either "multiple" or "single" for multiple or single row selection. Default is NULL.}

\item{showSortIcon}{Show a sort icon when sorting columns? Default is TRUE.}

\item{showSortable}{Show an indicator on sortable columns? Default is TRUE.}

\item{height}{Height of the table in pixels. Default is "auto" for automatic sizing.}

\item{width}{Width of the table in pixels. Default is "auto" for automatic sizing.}

\item{fullWidth}{Stretch the table to fill the full width of its container? Default is TRUE.}

\item{wrap}{Enable text wrapping? Default is TRUE.}

\item{outlined}{Add borders around the table? Default is FALSE.}

\item{bordered}{Add borders around the table and every cell? Default is FALSE.}

\item{borderless}{Remove inner borders from table? Default is TRUE.}

\item{striped}{Add zebra-striping to table rows? Default is TRUE.}

\item{compact}{Make tables more compact? Default is FALSE.}

\item{highlight}{Highlight table rows on hover? Default is FALSE.}

\item{pagination}{Enable pagination? Default is TRUE.}

\item{showPagination}{Show pagination? Default is NULL (defaults to TRUE if the table has more than one page).}

\item{showPageInfo}{Show page info? Default is TRUE.}

\item{showPageSizeOptions}{Show page size options? Default is FALSE.}

\item{paginationType}{Pagination control to use. Either "numbers" for page number buttons (the default),
"jump" for a page jump, or "simple" to show 'Previous' and 'Next' buttons only.}

\item{defaultPageSize}{Default page size for the table. Default is 10.}

\item{pageSizeOptions}{Page size options for the table. Default is 10, 25, 50, 100.}

\item{linksAsHyperlinks}{Cells that start with http, https or www will be converted to hyperlinks. Default is TRUE.}

\item{allow_point}{Boolean. If TRUE, allow this series' points to be selected by clicking on the graphic. Default is FALSE.}

\item{color_click}{String with color to change the color of the series that was clicked.}

\item{color_hover}{String with color to change the color of the series over which the mouse hovered.}

\item{cursor}{You can set the cursor to "pointer" if you have click events attached to the series, to signal to the user that the points and lines can be clicked.}

\item{clickFunction}{String with JS function to indicate the values to be captured when clicking on the chart.}

\item{dataLabels_show}{Boolean. If TRUE it shows the values of the chart labels}

\item{dataLabels_format_sample}{String. A specifier to format label(s)}

\item{dataLabels_size}{Number. Label font size}

\item{dataLabels_color}{String with hexadecimal color for the label}

\item{dataLabels_text_outline}{Boolean. If TRUE, data labels are show with outline}

\item{legend_decreasing}{Leaflet legend, If True decreasing order}

\item{opts}{All of the previous options can be passed as list of options opts}
}
\description{
This page describes the chart options arguments

The \code{tableOpts} options are used to style tables from the \code{reactable} package
and their descriptions have been copied and slightly adapted from their Github
\href{https://glin.github.io/reactable/reference/reactable.html}{documentation}.
}
