% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{default_theme_opts}
\alias{default_theme_opts}
\title{Theme Options}
\usage{
default_theme_opts(
  logo = NULL,
  logo_position = NULL,
  palette_colors = NULL,
  palette_colors_categorical = NULL,
  palette_colors_sequential = NULL,
  palette_colors_divergent = NULL,
  palette_type = NULL,
  branding_include = NULL,
  branding_text = NULL,
  background_color = NULL,
  branding_background_color = NULL,
  accent_color = NULL,
  text_show = NULL,
  text_size = NULL,
  text_color = NULL,
  text_family = NULL,
  line_color = NULL,
  line_size = NULL,
  title_color = NULL,
  title_align = NULL,
  title_family = NULL,
  title_size = NULL,
  title_weight = NULL,
  subtitle_color = NULL,
  subtitle_align = NULL,
  subtitle_size = NULL,
  subtitle_weight = NULL,
  caption_color = NULL,
  caption_align = NULL,
  caption_size = NULL,
  table_title_color = NULL,
  table_title_family = NULL,
  table_title_size = NULL,
  table_title_weight = NULL,
  table_subtitle_color = NULL,
  table_subtitle_family = NULL,
  table_subtitle_size = NULL,
  table_subtitle_weight = NULL,
  table_caption_color = NULL,
  table_caption_family = NULL,
  table_caption_size = NULL,
  table_caption_weight = NULL,
  header_text_color = NULL,
  header_text_family = NULL,
  header_text_size = NULL,
  header_text_weight = NULL,
  cell_text_color = NULL,
  cell_text_family = NULL,
  cell_text_size = NULL,
  cell_text_weight = NULL,
  link_color = NULL,
  header_background = NULL,
  table_background = NULL,
  table_border_color = NULL,
  zebra_stripe_color = NULL,
  desactivated_controls = NULL,
  axis_title_color = NULL,
  axis_title_size = NULL,
  axis_line_color = NULL,
  axis_line_size = NULL,
  axis_ticks_color = NULL,
  axis_tick_length = NULL,
  grid_x_enabled = NULL,
  grid_x_color = NULL,
  grid_x_size = NULL,
  grid_y_enabled = NULL,
  grid_y_color = NULL,
  grid_y_size = NULL,
  grid_y_dash = NULL,
  plot_margin_bottom = NULL,
  plot_margin_top = NULL,
  plot_margin_left = NULL,
  plot_margin_right = NULL,
  plot_background_color = NULL,
  plot_border_color = NULL,
  plot_border_size = NULL,
  label_color = NULL,
  label_size = NULL,
  label_text_decoration = NULL,
  label_text_outline = NULL,
  marker_enabled = TRUE
)
}
\description{
Theme Options
}
