% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-summarize.R
\name{preprocessData}
\alias{preprocessData}
\title{Drop NA}
\usage{
preprocessData(d, drop_na = FALSE)
}
\arguments{
\item{d}{A data frame}

\item{drop_na}{A logical value indicating if remove NA values}
}
\description{
Drop NA
}
